refactor: extract CreateIdentityTrait
