+++ src/Authentication/Actions/Email2FA.php
@@ -7,7 +7,6 @@ use CodeIgniter\HTTP\RedirectResponse;
 use CodeIgniter\Shield\Authentication\Authenticators\Session;
 use CodeIgniter\Shield\Entities\User;
 use CodeIgniter\Shield\Exceptions\RuntimeException;
-use CodeIgniter\Shield\Models\DatabaseException;
 use CodeIgniter\Shield\Models\UserIdentityModel;
 
 /**
@@ -17,6 +16,8 @@ use CodeIgniter\Shield\Models\UserIdentityModel;
  */
 class Email2FA implements ActionInterface
 {
+    use CreateIdentityTrait;
+
     private string $type = 'email_2fa';
 
     /**
@@ -114,52 +115,6 @@ class Email2FA implements ActionInterface
     public function afterLogin(User $user): void
     {
         $this->createIdentity($user, 'login', lang('Auth.need2FA'));
-    }
-
-    /**
-     * Create an identity for Email 2FA
-     */
-    private function createIdentity(User $user, string $name, string $extra): string
-    {
-        helper('text');
-
-        /** @var UserIdentityModel $userIdentityModel */
-        $userIdentityModel = model(UserIdentityModel::class);
-
-        // Delete any previous activation identities
-        $userIdentityModel->deleteIdentitiesByType($user, $this->type);
-
-        // Create an identity for our 2fa hash
-        $maxTry = 5;
-        $data   = [
-            'user_id' => $user->getAuthId(),
-            'type'    => $this->type,
-            'name'    => $name,
-            'extra'   => $extra,
-        ];
-
-        while (true) {
-            $data['secret'] = $this->generateSecretCode();
-
-            try {
-                $userIdentityModel->create($data);
-
-                break;
-            } catch (DatabaseException $e) {
-                $maxTry--;
-
-                if ($maxTry === 0) {
-                    throw $e;
-                }
-            }
-        }
-
-        return $data['secret'];
-    }
-
-    private function generateSecretCode(): string
-    {
-        return random_string('nozero', 6);
     }
 
     public function getType(): string
